pyproject.toml
[tool.poetry]
name = "moerphous-server"
version = "0.1.0"
description = "A Fully Async-based backend for Moerphous built using FastAPI, ODMantic, MongoDB, IPFS, XRPL-PY and friends."
authors = ["wiseaidev <business@wiseai.dev>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/moerphous/moerphous-server"
homepage = "https://github.com/moerphous/moerphous-server"
documentation = ""
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
]
keywords = ["python", "fastapi", "MongoDB", "ODMantic", "NFT Marketplace", "Moerphous"]
packages = [
    { include = "app" },
    { include = "LICENSE" },
    { include = "README.md" },
]

[tool.poetry.dependencies]
python = "^3.9.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi = {extras = ["uvicorn"], version = "^0.87.0"}
pyjwt = "^2.6.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
pydantic = {extras = ["email"], version = "^1.10.2"}
python-multipart = "^0.0.5"
odmantic = "^0.9.1"
xrpl-py = "^1.7.0"
pinatapy-vourhey = "^0.1.8"
dnspython = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "app.main:serve"
